--- rpc/jsonrpc/debug_block_access_list.go
@@ -1,34 +1,0 @@
-// Copyright 2026 The Erigon Authors
-// This file is part of Erigon.
-//
-// Erigon is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Erigon is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with Erigon. If not, see <http://www.gnu.org/licenses/>.
-
-package jsonrpc
-
-import (
-	"context"
-	"errors"
-
-	"github.com/erigontech/erigon/common/hexutil"
-	"github.com/erigontech/erigon/rpc"
-)
-
-// GetRawBlockAccessList returns the RLP-encoded block access list for a given block (EIP-7928).
-func (api *DebugAPIImpl) GetRawBlockAccessList(ctx context.Context, numberOrHash rpc.BlockNumberOrHash) (hexutil.Bytes, error) {
-	data, err := api.blockAccessListBytes(ctx, api.db, numberOrHash)
-	if errors.Is(err, errBlockAccessListNotFound) {
-		return nil, blockAccessListResourceNotFoundError()
-	}
-	return data, err
-}
